First day of Work
First day of working on KaTHUNK

Things are already starting to shape up, but nothing actually works and
there's TONS left to do. But it'll give me something to work on for the
next few weeks at least.

Currently the biggest blocker is UPNP since I really want that to work,
but I can never seem to get it to be consistent.