Second day of Work
Removed the UPNP code since I'm going to eventually switch it to WebRTC
or WebSockets for WASM support.

It can now load in a display the map, and you can move the tank around.
Currently rotating the barrel is wonky, so that needs to be fixed.

Started working on the code for joining games, but didn't finish it.